src/requirements.txt
streamlit
azure-ai-documentintelligence
streamlit
azure-storeage-blob
python-dotenv
azure.ai.documentintelligence
pandas